pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pygantry"
version = "1.2.0"
description = "Lightweight Python environment wrapper (Docker alternative using venv)"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [{name = "FBF / erabytse", email = "support@docudeeper.com"}]
dependencies = ["typer", "pyyaml", "rich"]

[project.scripts]
pygantry = "pygantry.cli:app"
pyg = "pygantry.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"